move utils/ under src/
