+++ src/avatar.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import TestUtils from 'react-addons-test-utils';
 import Avatar from './Avatar';
-import { variantTest } from '../utils/testUtils';
+import { variantTest } from './utils/testUtils';
 
 describe('Avatar', function() {
